GEODE-9788 pubsub benchmark -- improve logging and rename some classes
